+++ packages/vue/src/components/FCarousel/index.ts
@@ -1,0 +1,2 @@
+export * from './FCarousel'
+export * from './FCarouselItem'